requirements.txt
tweepy
discord.py
python-dateutil
async_lru